Break embed build scripts into modules and combine with existing build.js
